- Fixed most problems - Restructured some part of the code - Added built-in recursive operators (although disabled for now) - Resolved most of the conversion problems

+++ kernel/core/ParserLib/muParserTemplateMagic.h
@@ -4,7 +4,7 @@
 #include <cmath>
 #include "muParserError.h"
 
-std::complex<double> intPower(const std::complex<double>&, int);
+std::complex<double> intPower(const std::complex<double>&, int64_t);
 
 namespace mu
 {
